platform/apps/data-quality-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
clickhouse-driver==0.2.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
scipy==1.13.1
kafka-python==2.0.2
openlineage-python==1.20.0
pyyaml==6.0.2

